Improves performance of PriorityQueue.initialize()
By using a raw memory buffer for the '_order' array (which is static, once
created), we can effective initialize an array with all-zeroes much faster
than if we were using Swift's Array<Int>, since initializing such an array
with 'n' zeroes incurs a memcpy of every 4-to-8-byte-long Int, which is
inefficient.

This implementations improves performance of PriorityQueue.initialize()
by up to 10x.